feat: move get_markets to GammaClient
Move get_markets method from ClobClient to GammaClient
- Add GammaClient field to ClobClient for delegation
- Maintain backward compatibility via ClobClient::get_markets

Refs: #1
